path_hierarchy flattened and bug fixes
- Path hierarchy is a multi bucket aggregation so it should only return
a list of buckets.
Actually the response was build as each level of depth was a sub
aggregation repeating the name of the path_hierarchy aggregation.
IMO this is wrong, each level of depth in aggregation response should be
a real sub aggregation (like sum).
That's why I have flattened the response (see readme).
- Because of the previous point The hierarchy was stored in a Map (key =
path, value = list of buckets for that level) and
InternalPathHierarchy.getBuckets in InternalPathHierarchy was casting a
Map in a List (??) so we can't read the response with the java API.
- Bug on path_hierarchy with multivalues (values was override in
HierarchyValues.setDocument due to bad index)